# docs/openapi/charge-console-v1.5.3.yaml
openapi: 3.0.3
info:
  title: Charge Console API
  version: 1.5.3
  description: |
    Charge Console 1.5.3 添加前绑定引导、扫码确认与现有整桩提醒契约。
    POST /api/piles 沿用现有字段；YYB_BINDING_REQUIRED 和 YYB_RESCAN_REQUIRED 返回 409，可就地引导绑定。
    临时提醒仅在用户指定时长内按桩号低频检查，首次发现空闲并生成通知后自动结束。
    WxPusher 是附加发送方式；通知中心始终保存消息，平台成功发送不代表特定客户端已读。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
  /api/session/yyb-binding:
    get:
      operationId: getYybBinding
      summary: 读取当前用户的脱敏绑定与扫码能力，不刷新设备
      responses:
        "200":
          description: scanEnabled 仅表示 YYB 与 Mocele 均已配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YybBinding"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    delete:
      operationId: deleteYybBinding
      summary: 解除当前用户绑定并使其短期扫码会话失效（同账户连接操作进行中返回 409 YYB_BINDING_BUSY）
      responses:
        "204":
          description: 已解除
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/session/yyb-qr:
    post:
      operationId: createYybQr
      summary: 创建归属当前用户的短期扫码会话
      responses:
        "200":
          description: 后端重启或 30 分钟后需重新生成；不保存绑定
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YybQr"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          description: 未配置扫码或暂时无法创建会话
...
  /api/watch-rules:
    get:
      operationId: listWatchRules
      summary: 获取当前用户的全部整桩空闲提醒规则
      tags: [Watch Rules]
      responses:
        "200":
          description: 按更新时间倒序返回的空闲提醒规则
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchRule"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WatchUnavailable"
    post:
      operationId: createWatchRule
      summary: 为充电桩创建空闲提醒规则
      tags: [Watch Rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchRuleCreateRequest"
      responses:
        "201":
          description: 首次整桩检查结果；无空闲口时包含已创建的临时提醒
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchRuleCreateResult"
        "400":
          $ref: "#/components/responses/WatchInvalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "404":
          $ref: "#/components/responses/WatchTargetNotFound"
        "409":
          $ref: "#/components/responses/WatchConflict"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WatchUnavailable"
...
  /api/watch-overview:
    get:
      operationId: getWatchOverview
      summary: 获取当前用户的空闲提醒额度与后台策略摘要
      tags: [Watch Rules]
      responses:
        "200":
          description: 仅包含当前用户用量和管理员公开给普通用户的提醒策略
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchOverview"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WatchUnavailable"
  /api/notification-preferences:
    get:
      operationId: getNotificationPreference
      summary: 获取当前用户的通知偏好
      tags: [Notifications]
      responses:
        "200":
          description: 浏览器通知意愿与免打扰设置
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreference"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/NotificationUnavailable"
    patch:
      operationId: updateNotificationPreference
      summary: 更新通知偏好
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferenceUpdateRequest"
      responses:
        "200":
          description: 更新后的通知偏好
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreference"
        "400":
          $ref: "#/components/responses/NotificationPreferenceInvalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/NotificationUnavailable"
  /api/notifications:
    get:
      operationId: listNotifications
      summary: 分页获取当前用户的通知中心消息
      tags: [Notifications]
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NotificationStatusFilter"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 通知分页、下一游标和总未读数量
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPage"
        "400":
          $ref: "#/components/responses/NotificationQueryInvalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/NotificationUnavailable"
...
  /api/notification-channels/wxpusher:
    get:
      operationId: getWxPusherChannel
      summary: 获取当前用户的 WxPusher 通道状态
      tags: [WxPusher]
      responses:
        "200":
          description: 配置可用性、绑定状态、事件偏好和最近投递结果；不返回完整 UID 或 AppToken
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WxPusherChannelState"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WxPusherUnavailable"
    patch:
      operationId: updateWxPusherChannel
      summary: 更新 WxPusher 总开关和事件类型偏好
      tags: [WxPusher]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WxPusherChannelUpdateRequest"
      responses:
        "200":
          description: 更新后的通道状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WxPusherChannelState"
        "400":
          $ref: "#/components/responses/WxPusherInvalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "409":
          $ref: "#/components/responses/WxPusherChannelConflict"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WxPusherUnavailable"
    delete:
      operationId: deleteWxPusherChannel
      summary: 解除当前用户的本地 WxPusher 绑定
      description: 删除本地 UID 并取消尚未开始的投递任务，不承诺替用户取消关注 WxPusher 应用或退出接收客户端。
      tags: [WxPusher]
      responses:
        "204":
          description: 已解除绑定；重复调用同样成功
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WxPusherUnavailable"
  /api/notification-channels/wxpusher/bind-sessions:
    post:
      operationId: createWxPusherBindSession
      summary: 创建十分钟有效的 WxPusher 参数二维码绑定会话
      tags: [WxPusher]
      responses:
        "201":
          description: 等待扫码的绑定会话；Provider code 不对前端公开
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WxPusherBindSession"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "409":
          $ref: "#/components/responses/WxPusherBindingConflict"
        "429":
          $ref: "#/components/responses/WxPusherRateLimited"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WxPusherUnavailable"
...
  /api/notification-channels/wxpusher/test:
    post:
      operationId: testWxPusherChannel
      summary: 为当前用户创建一条受频率限制的 WxPusher 测试投递
      tags: [WxPusher]
      responses:
        "202":
          description: 测试投递已进入持久队列
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationDeliverySummary"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "409":
          $ref: "#/components/responses/WxPusherChannelConflict"
        "429":
          $ref: "#/components/responses/WxPusherRateLimited"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WxPusherUnavailable"
  /api/notification-channels/wxpusher/test/recheck:
    post:
      operationId: recheckWxPusherTestDelivery
      summary: 重新查询当前用户最近一条 WxPusher 测试消息的供应商状态
      description: 仅查询原投递记录，不创建或重新发送消息。已结束或尚未提交到供应商的状态会原样返回。
      tags: [WxPusher]
      responses:
        "200":
          description: 重新查询后的测试投递状态
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationDeliverySummary"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/OrdinaryUserRequired"
        "404":
          description: 当前用户没有测试投递记录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/WxPusherUnavailable"
  /api/admin/trends:
    get:
      operationId: getAdminTrends
      summary: 获取管理员运营趋势
      tags: [Admin Trends]
      parameters:
        - $ref: "#/components/parameters/AdminTrendRange"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: 请求量、远端成功率、活跃用户和离线端口趋势
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTrendsResponse"
        "400":
          $ref: "#/components/responses/InvalidAdminTrendQuery"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AdminRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/AdminTrendsUnavailable"
  /api/admin/trends.csv:
    get:
      operationId: exportAdminTrendsCSV
      summary: 导出管理员运营趋势 CSV
      tags: [Admin Trends]
      parameters:
        - $ref: "#/components/parameters/AdminTrendRange"
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: 与相同范围趋势接口数值一致的 UTF-8 BOM CSV 文件
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
            Content-Disposition:
              $ref: "#/components/headers/CSVContentDisposition"
          content:
            text/csv:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/InvalidAdminTrendQuery"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AdminRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/AdminTrendsUnavailable"
  /api/admin/settings:
    get:
      operationId: getAdminSettings
      summary: 获取系统注册、提醒与保留策略
      tags: [Admin Operations]
      responses:
        "200":
          description: 当前完整系统策略
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationSettings"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AdminRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    put:
      operationId: updateAdminSettings
      summary: 更新系统注册、提醒与保留策略
      description: 修改结果会写入管理员审计日志，并立即唤醒提醒调度器。
      tags: [Admin Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationSettings"
      responses:
        "200":
          description: 保存后的完整系统策略
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationSettings"
        "400":
          description: 参数范围或时区无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AdminRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/admin/operations:
    get:
      operationId: getAdminOperations
      summary: 获取数据库、备份、通知与提醒调度状态
      tags: [Admin Operations]
      responses:
        "200":
          description: 当前运维状态和过去 24 小时提醒调度指标
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationsStatus"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AdminRequired"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          description: 数据库运维信息暂时不可用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: charge_session
...
  headers:
    PrivateNoStore:
      description: 分析响应包含用户或运营数据，不允许共享或本地缓存
      schema:
        type: string
        enum: ["private, no-store"]
...
  responses:
...
    Unauthenticated:
      description: 未登录或会话已失效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: login required
    OrdinaryUserRequired:
      description: 当前账户不是普通用户
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: ordinary user permission required
...
    MethodNotAllowed:
      description: 请求方法不受支持
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: method not allowed
...
    WatchUnavailable:
      description: 空闲提醒规则存储暂时不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodedErrorResponse"
          example:
            code: WATCH_UNAVAILABLE
            error: 空闲提醒功能暂时不可用，请稍后重试
...
  schemas:
    YybBinding:
      type: object
      required: [bound]
      properties:
        bound: { type: boolean }
        scanEnabled:
          type: boolean
          description: YYB 与 Mocele 已配置；旧接口可能缺少此字段
        sessionId:
          type: string
          description: 仅确认结果包含，用于证明本次二维码的绑定已保存；不得用旧 bound=true 推断新会话成功
        nickname: { type: string }
        openidSuffix: { type: string }
        status: { type: string }
        boundAt: { type: string, format: date-time }
        lastCheckedAt: { type: string, format: date-time }
        cookieSynced: { type: boolean }
        syncState:
          type: string
          enum: [not_needed, synced, failed]
          description: 无设备无需同步、同步成功、绑定已保存但同步失败
        message: { type: string }
    YybQr:
      type: object
      required: [sessionId]
      properties:
        sessionId: { type: string }
        imageUrl: { type: string }
        imageBase64: { type: string }
        status: { type: string }
...
    WatchRuleCreateRequest:
      type: object
      additionalProperties: false
      required: [deviceId]
      properties:
        deviceId:
          type: string
          pattern: "^[0-9]{6,64}$"
        duration:
          $ref: "#/components/schemas/WatchTemporaryDuration"
          description: 省略时使用 2h，且服务端会将截止时间限制在当天计划断电开始前。
...